refactored generate_waveform_and_duration to separate file in app/utils/waveform_gen.py
